requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
openai==1.12.0
python-dotenv==1.0.0
websockets==12.0
pydantic==2.6.1
pydantic-settings==2.2.1
pytest==8.0.2
pytest-asyncio==0.23.5